Disable legacy view support for VMs
 - Removed the routes for `/vm-legacy` and `/vm/{id}/edit` since
   both functions are now covered by `/vm`

 - Removed the pages that supported the Legacy VM details and edit

 - Adjusted the VM card list to open the `/vm` VM dashboard details
   page directly

 - Adjusted the actions to only support the current dashboard details
   page and to consistently render spacing between buttons
