+++ src/components/VmActions/Action.js
@@ -1,8 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-
-import Button from './Button'
 import { excludeKeys } from 'patternfly-react'
+
+import { hrefWithoutHistory } from '_/helpers'
+
+import style from './style.css'
 
 class Action extends React.Component {
   constructor (props) {
@@ -35,7 +37,6 @@ class Action extends React.Component {
     </React.Fragment>
   }
 }
-
 Action.propTypes = {
   children: PropTypes.node.isRequired,
   confirmation: PropTypes.node,
@@ -43,11 +44,53 @@ Action.propTypes = {
 
 export default Action
 
+class Button extends React.Component {
+  render () {
+    let {
+      className,
+      tooltip = '',
+      actionDisabled = false,
+      onClick,
+      shortTitle,
+      id,
+    } = this.props
+
+    let handleClick = hrefWithoutHistory(onClick)
+
+    if (actionDisabled) {
+      return (
+        <button className={`${className} ${style['disabled-button']}`} disabled='disabled' id={id}>
+          <span data-toggle='tooltip' data-placement='left' title={tooltip}>
+            {shortTitle}
+          </span>
+        </button>
+      )
+    }
+
+    return (
+      <span className={style['full-button']}>
+        <a href='#' onClick={handleClick} className={`${className} ${style['link']}`} id={id}>
+          <span data-toggle='tooltip' data-placement='left' title={tooltip} id={`${id}-title`}>
+            {shortTitle}
+          </span>
+        </a>
+      </span>
+    )
+  }
+}
+Button.propTypes = {
+  className: PropTypes.string.isRequired,
+  tooltip: PropTypes.string,
+  shortTitle: PropTypes.string.isRequired,
+  onClick: PropTypes.func,
+  actionDisabled: PropTypes.bool,
+  id: PropTypes.string.isRequired,
+}
+
 const ActionButtonWraper = (props) => {
   const btnProps = excludeKeys(props, [ 'confirmation' ])
   return <Action confirmation={props.confirmation} key={props.shortTitle}><Button {...btnProps} /></Action>
 }
-
 ActionButtonWraper.propTypes = {
   confirmation: PropTypes.node,
   ...Button.propTypes,
